finance_ingestion/requirements.txt
requests==2.31.0
python-dotenv==1.0.0
newspaper3k==0.2.8
lxml_html_clean==0.1.1
yfinance==0.2.40
transformers==4.41.2
torch==2.2.2
numpy==1.26.4
groq==0.4.1
httpx==0.27.0